+++ src/errors/database_error.rs
